requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
strawberry-graphql==0.218.0
strawberry-graphql-fastapi==0.10.0
